+++ docs/release_notes.rst
@@ -1,9 +1,16 @@
 Release Notes
 =============
 
+v1.9.0
+------
+* Add DataKitchenClient methods to get and delete orders in a kitchen.
+* Add DataKitchenClient method to get recipe contents.
+* Add functions for extracting tests and test metadata from recipes.
+
 v1.8.2
 ------
 * Fix a problem in the GMailClient that was causing excel spreadsheets sent as attachments to be corrupted.
+
 v1.8.1
 ------
 * Update setup.py and docs in preparation for move to public repository